# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v2/tenants/{tenantId}/devices/{deviceId}/policy": {
            "get": {
                "description": "获取指定设备当前应用的策略",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "获取设备策略",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SecurityPolicy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v2/tenants/{tenantId}/devices/{deviceId}/whitelist": {
            "get": {
                "description": "获取指定设备的应用白名单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "白名单管理"
                ],
                "summary": "获取设备白名单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/policies": {
            "get": {
                "description": "获取指定租户下的所有策略列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "列出租户策略",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "为租户创建新的安全策略",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "创建安全策略",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "创建策略请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreatePolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/policies/{policyId}": {
            "get": {
                "description": "根据策略ID获取策略详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "获取策略详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "策略ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SecurityPolicy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定策略的信息和设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "更新策略",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "策略ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新策略请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatePolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SecurityPolicy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的策略",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "删除策略",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "策略ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/policies/{policyId}/assign": {
            "post": {
                "description": "将策略分配给指定的设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "策略管理"
                ],
                "summary": "分配策略",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "策略ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分配策略请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AssignPolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/policies/{policyId}/whitelist": {
            "post": {
                "description": "将指定应用添加到策略的白名单中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "白名单管理"
                ],
                "summary": "添加应用到白名单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "策略ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "添加应用请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddAppRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/policies/{policyId}/whitelist/{packageName}": {
            "delete": {
                "description": "将指定应用从策略的白名单中移除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "白名单管理"
                ],
                "summary": "从白名单移除应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "策略ID",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "应用包名",
                        "name": "packageName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AddAppRequest": {
            "type": "object",
            "properties": {
                "allow_notifications": {
                    "type": "boolean"
                },
                "app_name": {
                    "type": "string"
                },
                "auto_launch": {
                    "type": "boolean"
                },
                "default_shortcut": {
                    "type": "boolean"
                },
                "package_name": {
                    "type": "string"
                }
            }
        },
        "api.AssignPolicyRequest": {
            "type": "object",
            "properties": {
                "device_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.CreatePolicyRequest": {
            "type": "object",
            "properties": {
                "app_whitelist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AppWhitelistEntry"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/models.SecurityPolicySettings"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "api.UpdatePolicyRequest": {
            "type": "object",
            "properties": {
                "app_whitelist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AppWhitelistEntry"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/models.SecurityPolicySettings"
                }
            }
        },
...
        "models.AppWhitelistEntry": {
            "type": "object",
            "properties": {
                "allow_notifications": {
                    "description": "允许通知",
                    "type": "boolean"
                },
                "app_name": {
                    "description": "应用名称",
                    "type": "string"
                },
                "auto_launch": {
                    "description": "自启动",
                    "type": "boolean"
                },
                "default_shortcut": {
                    "description": "显示默认快捷方式",
                    "type": "boolean"
                },
                "package_name": {
                    "description": "包名",
                    "type": "string"
                }
            }
        },
...
        "models.SecurityPolicy": {
            "type": "object",
            "properties": {
                "app_whitelist": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AppWhitelistEntry"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/models.SecurityPolicySettings"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}